extracted out ts component & state files
