+++ index.html
@@ -100,12 +100,6 @@
         </div>
         
         <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-        <script src="assets/js/templateClass.js"></script>
-        <script src="assets/js/stateClass.js"></script>
-        <script src="assets/js/gameState.js"></script>
-        <script src="assets/js/engrams.js"></script>
-        <script src="assets/js/scoreboard.js"></script>
-        <script src="assets/js/app.js"></script>
         <script src="dist/bundle.js"></script>
     </body>
 </html>
